llext: add a section reordering script
In some cases, the compiler generates objects that have sections
"out of order" for the LLEXT loader, i.e. regions overlap.
As ET_REL objects are not produced with the linker, a linker script
cannot be used to correct the section order. Reordering variable
declarations, code, header includes, etc. rarely fixes the issue
either. This script detects if the LLEXT loader will reject an
ET_REL object and reorders the sections to match LLEXT regions.

Signed-off-by: Lauren Murphy <lauren.murphy@intel.com>