llext: replace arcextmap shstrtab strip script with call to gnu strip
Use GNU strip to remove the .arcextmap sections (and their names from the
shstrtab) before calling stripac as an alternative to invoking custom
Python script that removes .arcextmap section names after stripac
removes the sections (but fails to remove their names).

Not sure this is advisable, but tests aren't failing.

Signed-off-by: Lauren Murphy <lauren.murphy@intel.com>